(sql-migration-connector) Start isolating quaint imports...
...in connection_wrapper module. To isolate the surface coupled to our
database drivers. This should also disentangle error handling. At the
lower level, we use quaint, but we don't refer to it anymore once we
have a user facing error error code.
